Moved gps_to_utc from process_vcs to general_utils
